backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.3.0"
description = ""
authors = ["Alexander Demure"]

[tool.poetry.dependencies]

python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.5"
structlog = "^20.1.0"
tenacity = "^6.2.0"
aiofiles = "^0.5.0"

httpx = "^0.16.1"
pytest-asyncio = "^0.14.0"

email-validator = "^1.1.1"
jinja2 ="^2.11.2"

itsdangerous = "^1.1.0"
python-jose = "^3.2.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^3.1.7"
boto3 = "^1.16.25"
yandex-checkout = "^1.6"

tortoise-orm = "^0.16.5"
psycopg2-binary = "^2.8.5"
aerich = "^0.2.1"
databases = "^0.3.2"
asyncpg = "^0.20.1"
aiosqlite = "^0.15.0"

aioredis = "^1.3.1"
ujson = "^4.0.1"

pillow = "^8.1.0"

aiocache = "^0.11.1"

sentry-sdk = "^0.19.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
